refactor: apple-music parser rename
